auto-assignment fixed for Payload attributes!! All I need to do now is make a method for the WasteBin Factory

+++ src/main/java/be/kdg/magiwastebackend/domain/RawDataLog.java
@@ -1,6 +1,5 @@
 package be.kdg.magiwastebackend.domain;
 
-import be.kdg.magiwastebackend.service.RawDataLogService;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -9,13 +8,15 @@ import jakarta.persistence.Id;
 @Entity
 public class RawDataLog {
 
-    float temperature;
+    double temperature;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private float sensorDistance1;
-    private float sensorDistance2;
+    //TODO add attribute for bin id
+
+    private double sensorDistance1;
+    private double sensorDistance2;
     private boolean tiltState;
     //this field stores the unused controller payload keys and values as a json string. yikes!
     private String unusedData;
@@ -37,7 +38,7 @@ public class RawDataLog {
 
     }
 
-    public RawDataLog(float sensorDistance1, float sensorDistance2, boolean tiltState, float temperature, String unusedDataJsonString) {
+    public RawDataLog(double sensorDistance1, double sensorDistance2, boolean tiltState, double temperature, String unusedDataJsonString) {
         this.sensorDistance1 = sensorDistance1;
         this.sensorDistance2 = sensorDistance2;
         this.tiltState = tiltState;
@@ -54,19 +55,19 @@ public class RawDataLog {
         return id;
     }
 
-    public float getSensorDistance1() {
+    public double getSensorDistance1() {
         return sensorDistance1;
     }
 
-    public void setSensorDistance1(float sensorDistance1) {
+    public void setSensorDistance1(double sensorDistance1) {
         this.sensorDistance1 = sensorDistance1;
     }
 
-    public float getSensorDistance2() {
+    public double getSensorDistance2() {
         return sensorDistance2;
     }
 
-    public void setSensorDistance2(float sensorDistance2) {
+    public void setSensorDistance2(double sensorDistance2) {
         this.sensorDistance2 = sensorDistance2;
     }
